Add new Map method to generate maps
add helper methods to keep style and keep the code dry

we needed to add a highlight method to hightlight the section of the
map.
since we're initiating the map with a default grey style we simply
cannot use the built-in method "resetStyle"
(http://leafletjs.com/reference.html#geojson-resetstyle) since this
would be the default gray instead of the calculated color.
Thats why we need to retain the style for each map. And since we have
two maps it needs to be an instance varaible for the map.

I opted to go with a class style method to keep everything in one place
and don't clutter the DOM (which would be another way to save the style)