Add small "hack" for feature highlighting
keep the highlight if a popup is opened (click on the map) but remove
the highlight if the popup is closed (close button or click somehwere on
the map)
this is accomplished by a state variable which keeps the state of the
popup beeing opened via a click on the map.
the "click" event is handled first, when a popup is currently open the
"popupclose" event is triggered which would remove the hightlight for
the feature but since we set the variable/flag that it was opened via
the map click event we do not remove the highlight but set the flat to
false. we need the "popupopen" event because when there is no popup the
"popupclose" event is not called and the variable/flag is not reset.
therefore we need the third event to set the flag to false, to allow the
hightlight to be removed once the user closes the popup or clicks
somehwere on the map